+++ src/util.rs
@@ -1,13 +1,10 @@
 // Copyright 2023 the Vello Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use vello::Scene;
-use vello::kurbo::{Affine, BezPath, Point, Rect, Stroke};
-use vello::peniko::color::{DynamicColor, palette};
-use vello::peniko::{Brush, Color, Fill};
-
-#[cfg(feature = "image")]
-use vello::peniko::{Blob, ImageBrush};
+use crate::RenderSink;
+use kurbo::{Affine, BezPath, Cap, Join, Point, Rect, Stroke};
+use peniko::color::{DynamicColor, palette};
+use peniko::{Brush, Color, ColorStop, Fill, Gradient};
 
 pub fn to_affine(ts: &usvg::Transform) -> Affine {
     let usvg::Transform {
@@ -24,14 +21,14 @@ pub fn to_affine(ts: &usvg::Transform) -> Affine {
 pub fn to_stroke(stroke: &usvg::Stroke) -> Stroke {
     let mut conv_stroke = Stroke::new(stroke.width().get() as f64)
         .with_caps(match stroke.linecap() {
-            usvg::LineCap::Butt => vello::kurbo::Cap::Butt,
-            usvg::LineCap::Round => vello::kurbo::Cap::Round,
-            usvg::LineCap::Square => vello::kurbo::Cap::Square,
+            usvg::LineCap::Butt => Cap::Butt,
+            usvg::LineCap::Round => Cap::Round,
+            usvg::LineCap::Square => Cap::Square,
         })
         .with_join(match stroke.linejoin() {
-            usvg::LineJoin::Miter | usvg::LineJoin::MiterClip => vello::kurbo::Join::Miter,
-            usvg::LineJoin::Round => vello::kurbo::Join::Round,
-            usvg::LineJoin::Bevel => vello::kurbo::Join::Bevel,
+            usvg::LineJoin::Miter | usvg::LineJoin::MiterClip => Join::Miter,
+            usvg::LineJoin::Round => Join::Round,
+            usvg::LineJoin::Bevel => Join::Bevel,
         })
         .with_miter_limit(stroke.miterlimit().get() as f64);
     if let Some(dash_array) = stroke.dasharray().as_ref() {
@@ -93,23 +90,6 @@ pub fn to_bez_path(path: &usvg::Path) -> BezPath {
     local_path
 }
 
-#[cfg(feature = "image")]
-pub fn into_image(image: image::ImageBuffer<image::Rgba<u8>, Vec<u8>>) -> ImageBrush {
-    use vello::peniko::ImageAlphaType;
-    use vello::peniko::ImageData;
-
-    let (width, height) = (image.width(), image.height());
-    let image_data: Vec<u8> = image.into_vec();
-    ImageData {
-        data: Blob::new(std::sync::Arc::new(image_data)),
-        format: vello::peniko::ImageFormat::Rgba8,
-        alpha_type: ImageAlphaType::AlphaPremultiplied,
-        width,
-        height,
-    }
-    .into()
-}
-
 pub fn to_brush(paint: &usvg::Paint, opacity: usvg::Opacity) -> Option<(Brush, Affine)> {
     match paint {
         usvg::Paint::Color(color) => Some((
@@ -122,10 +102,10 @@ pub fn to_brush(paint: &usvg::Paint, opacity: usvg::Opacity) -> Option<(Brush, A
             Affine::IDENTITY,
         )),
         usvg::Paint::LinearGradient(gr) => {
-            let stops: Vec<vello::peniko::ColorStop> = gr
+            let stops: Vec<ColorStop> = gr
                 .stops()
                 .iter()
-                .map(|stop| vello::peniko::ColorStop {
+                .map(|stop| ColorStop {
                     offset: stop.offset().get(),
                     color: DynamicColor::from_alpha_color(Color::from_rgba8(
                         stop.color().red,
@@ -147,15 +127,14 @@ pub fn to_brush(paint: &usvg::Paint, opacity: usvg::Opacity) -> Option<(Brush, A
             ]
             .map(f64::from);
             let transform = Affine::new(arr);
-            let gradient =
-                vello::peniko::Gradient::new_linear(start, end).with_stops(stops.as_slice());
+            let gradient = Gradient::new_linear(start, end).with_stops(stops.as_slice());
             Some((Brush::Gradient(gradient), transform))
         }
         usvg::Paint::RadialGradient(gr) => {
-            let stops: Vec<vello::peniko::ColorStop> = gr
+            let stops: Vec<ColorStop> = gr
                 .stops()
                 .iter()
-                .map(|stop| vello::peniko::ColorStop {
+                .map(|stop| ColorStop {
                     offset: stop.offset().get(),
                     color: DynamicColor::from_alpha_color(Color::from_rgba8(
                         stop.color().red,
@@ -180,13 +159,9 @@ pub fn to_brush(paint: &usvg::Paint, opacity: usvg::Opacity) -> Option<(Brush, A
             ]
             .map(f64::from);
             let transform = Affine::new(arr);
-            let gradient = vello::peniko::Gradient::new_two_point_radial(
-                start_center,
-                start_radius,
-                end_center,
-                end_radius,
-            )
-            .with_stops(stops.as_slice());
+            let gradient =
+                Gradient::new_two_point_radial(start_center, start_radius, end_center, end_radius)
+                    .with_stops(stops.as_slice());
             Some((Brush::Gradient(gradient), transform))
         }
         usvg::Paint::Pattern(_) => None,
@@ -195,7 +170,7 @@ pub fn to_brush(paint: &usvg::Paint, opacity: usvg::Opacity) -> Option<(Brush, A
 
 /// Error handler function for [`super::append_tree_with`] which draws a transparent red box
 /// instead of unsupported SVG features
-pub fn default_error_handler(scene: &mut Scene, node: &usvg::Node) {
+pub fn default_error_handler(scene: &mut impl RenderSink, node: &usvg::Node) {
     let bb = node.bounding_box();
     let rect = Rect {
         x0: bb.left() as f64,
@@ -206,27 +181,8 @@ pub fn default_error_handler(scene: &mut Scene, node: &usvg::Node) {
     scene.fill(
         Fill::NonZero,
         Affine::IDENTITY,
-        palette::css::RED.with_alpha(0.5),
-        None,
+        &Brush::Solid(palette::css::RED.with_alpha(0.5)),
+        Affine::IDENTITY,
         &rect,
     );
-}
-
-#[cfg(feature = "image")]
-pub fn decode_raw_raster_image(
-    img: &usvg::ImageKind,
-) -> Result<image::RgbaImage, image::ImageError> {
-    // All `image::ImageFormat` variants exist even if the feature in the image crate is disabled,
-    // but `image::load_from_memory_with_format` will fail with an Unsupported error if the
-    // image crate feature flag is disabled. So we don't need any of our own feature handling here.
-    let (data, format) = match img {
-        usvg::ImageKind::JPEG(data) => (data, image::ImageFormat::Jpeg),
-        usvg::ImageKind::PNG(data) => (data, image::ImageFormat::Png),
-        usvg::ImageKind::GIF(data) => (data, image::ImageFormat::Gif),
-        usvg::ImageKind::WEBP(data) => (data, image::ImageFormat::WebP),
-        usvg::ImageKind::SVG(_) => unreachable!(),
-    };
-
-    let dyn_image = image::load_from_memory_with_format(data, format)?;
-    Ok(dyn_image.into_rgba8())
 }