+++ CHANGELOG.md
@@ -15,6 +15,15 @@ You can find its changes [documented below](#090-2026-01-29).
 
 This release has an [MSRV][] of 1.88.
 
+### Added
+
+- Added the `RenderSink` abstraction and generic `append*` APIs ([#100][] by [@RobertBrewitz])
+
+### Changed
+
+- **Breaking:** Removed the `image` and `image_format_*` features and the public helpers `into_image` and `decode_raw_raster_image`, implement `RenderSink::draw_image` to handle decoding and rendering.
+- Made Vello optional behind the default `vello` feature ([#100][] by [@RobertBrewitz])
+
 ## [0.10.0][] (2026-07-19)
 
 This release has an [MSRV][] of 1.88.
@@ -161,7 +170,6 @@ This release has an [MSRV][] of 1.75.
 [@MarijnS95]: https://github.com/MarijnS95
 [@DasLixou]: https://github.com/DasLixou
 [@RobertBrewitz]: https://github.com/RobertBrewitz
-
 [#85]: https://github.com/linebender/vello_svg/pull/85
 [#63]: https://github.com/linebender/vello_svg/pull/63
 [#61]: https://github.com/linebender/vello_svg/pull/61
@@ -179,7 +187,6 @@ This release has an [MSRV][] of 1.75.
 [#8]: https://github.com/linebender/vello_svg/pull/8
 [#6]: https://github.com/linebender/vello_svg/pull/6
 [#1]: https://github.com/linebender/vello_svg/pull/1
-
 [Unreleased]: https://github.com/linebender/vello_svg/compare/v0.10.0...HEAD
 [0.10.0]: https://github.com/linebender/vello_svg/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/linebender/vello_svg/compare/v0.8.0...v0.9.0
@@ -193,5 +200,4 @@ This release has an [MSRV][] of 1.75.
 [0.3.0]: https://github.com/linebender/vello_svg/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/linebender/vello_svg/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/linebender/vello_svg/releases/tag/v0.1.0
-
 [MSRV]: README.md#minimum-supported-rust-version-msrv